Remove chapter V title, duplicate of VI
The titles of chapter V and VI were duplicates of each other.

It seems that chapter VI's title is more befitting than that of chapter V.

Leave chapter V untitled for now. Chapter I is also untitled, so there
is a precedent. In later work, perhaps titles will be suggested.