Transliterate to US ASCII for portability
Convert from UTF8 enconding to US-ASCII for portability. In particular
it was discovered that pandoc's markdown to PDF conversion is not keen
on the previous encoding by default.

Converted like this:

    $ iconv -f utf8 -t us-ascii//translit barrie_bear.md

The end result seems to have expanded ellipses, converted 'smart quotes'
back to the regular variety, and converted hypens. One instance of the
word 'flood' had an interesting conversion, perhaps a funky unicode
substitution was used during OCR.