Move title to yaml frontmatter, as book
Move the title from a heading 1, to an attribute in the yaml
frontmatter. Set the metadata to specify that this is an A5 book.

When rendering this document as a book using pandoc, better results are
achieved if we specify the title in the metadata and specify what kind
of document this is.